Move FundingCycle/ => components/v1/
