+++ src/components/v1/FundingCycle/ReservedTokens.tsx
@@ -14,7 +14,7 @@ import { decodeFundingCycleMetadata } from 'utils/fundingCycle'
 import { readNetwork } from 'constants/networks'
 
 import DistributeTokensModal from './modals/DistributeTokensModal'
-import TicketModsList from '../shared/TicketModsList'
+import TicketModsList from '../../shared/TicketModsList'
 import { V1_PROJECT_IDS } from 'constants/v1/projectIds'
 
 export default function ReservedTokens({
