+++ src/components/v1/V1Project/FundingCycles/FundingHistory.tsx
@@ -17,7 +17,7 @@ import { formatHistoricalDate } from 'utils/formatDate'
 import { formatWad } from 'utils/formatNumber'
 import { hasFundingTarget } from 'utils/fundingCycle'
 
-import FundingCycleDetails from 'components/FundingCycle/FundingCycleDetails'
+import FundingCycleDetails from 'components/v1/FundingCycle/FundingCycleDetails'
 
 export default function FundingHistory({
   startId,
